# Performance Improvement: Only execute the computationally expensive t…