…hermal strength calculation every x-th iteration.

_Motviation_:
On my PC (4-core Xeon 2.8Ghz, 16GB RAM, Quadro M2000m graphics card), the X-Plane framerate drops from ~100fps to 30-60FPS when activating the xplane-thermals module. There is probably other reasons and ways to improve performance, but I am implementing one initial/simple method here.

_Method_:
Super simple method, just do the calculations for actual thermal lift and roll values once every x-th iteration (Specified by the `performance_divider` variable), but still provide the "saved" values from the last calculation on every single iteration. For me, a value of 5-10 gives a significant performance boost (~80fps) while not having any noticable effect on the thermal behavior (I am checking this via X-Planes "Show Flight Model" Option). Note that even when just calculated every x-th frame, at 100FPS the thermals would still be calculated with 10Hz, which should be more than enough.

Note: @alexferrer I noticed that I was using `return -1` here https://github.com/alexferrer/xplane_thermals/blob/master/PI_ThermalSim.py#L240, while you are using `return 0.04` I guess in your case, the additional thermal lift values are then actually not applied every single frame? I.e. if you have for example 100FPS but run the `FlightLoopCallback` function only ever 0.04 seconds, then the additional lift value would jump in the sort 0...0...0...X...0...0...0...Y...0...0...0...Z... . That is why I though it is better to store the lift values when they are recalculated, but supply them every time (lift values then look like X...X...X...X...Y...Y...Y...Y...Z...Z...Z...Z...).

_Status_: 
Ready to merge.