@alexferrer 2016-06-27
The way I understand the plugin, the plugin is executed once per FlightLoopCallback iteration. nothing happens in between.

REF: http://www.xsquawkbox.net/xpsdk/mediawiki/Category:XPLMProcessing

With .04 it gets called (and it "injects" energy to the plane) every .04 second,  For  a glider flying at 100kph that would be ~ 1 meter change of position. 

If I am reading it right, your change loops every frame (value -1) but test position and applies energy every 5. If that is the case, would it not be easier to change the value to -5 (Every 5th frame) ?

Alex.