@philipoe 2016-06-27
> If I am reading it right, your change loops every frame (value -1) but test position and applies energy every 5?

Yes, this is the case for my loop. 

>  If that is the case, would it not be easier to change the value to -5 (Every 5th frame) 

> The way I understand the plugin, the plugin is executed once per FlightLoopCallback iteration. nothing happens in between.

I honestly only very quickly browsed through that documentation, but I cannot determine exactly what happens "in between" from this documentaiton. However, when I just pass "-5" as a return value, then this does _not_ work. Try it by returning something like -200 (call loop every 2 seconds on a 100FPS machine). You will see that the lift values applied by the xplanethermals-plugin peak, then for 2 seconds they are at zero, and then they peak again. This also makes kind of sense to me, as we are writing the `self.roll.value` and other values only when the flight loop callback is actually executed.